requirements.txt
fastapi>=0.124.1
uvicorn>=0.31.0
pydantic[email]>=2.12.0
prisma==0.15.0
pyjwt>=2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiohttp==3.10.8
python-dotenv==1.0.1
google-genai>=1.56.0
google-adk==1.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
notion-client==2.2.1
PyGithub==2.6.1
GitPython==3.1.44
anthropic[vertex]>=0.30.0
google-cloud-billing>=1.15.0
google-cloud-monitoring>=2.22.0
google-cloud-bigquery>=3.25.0
apscheduler==3.10.4
